+++ src/serviceUnavailableInterceptor/serviceUnavailableInterceptor.ts
@@ -1,10 +1,10 @@
-import {Injectable, Optional, Injector, ClassProvider} from "@angular/core";
-import {HttpInterceptor, HttpHandler, HttpEvent, HTTP_INTERCEPTORS, HttpRequest} from "@angular/common/http";
-import {IgnoredInterceptorsService, IgnoredInterceptorId, AdditionalInfo} from "@anglr/common";
-import {Observable, Observer} from "rxjs";
-import {catchError} from "rxjs/operators";
+import {Injectable, Optional, Injector, ClassProvider} from '@angular/core';
+import {HttpInterceptor, HttpHandler, HttpEvent, HTTP_INTERCEPTORS, HttpRequest} from '@angular/common/http';
+import {IgnoredInterceptorsService, IgnoredInterceptorId, AdditionalInfo} from '@anglr/common';
+import {Observable, Observer} from 'rxjs';
+import {catchError} from 'rxjs/operators';
 
-import {ServiceUnavailableInterceptorOptions} from "./serviceUnavailableInterceptorOptions";
+import {ServiceUnavailableInterceptorOptions} from './serviceUnavailableInterceptorOptions';
 
 /**
  * ServiceUnavailableInterceptor used for intercepting http responses and handling 503 statuses