+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Version 8.0.0 (2020-08-25)
+## Version 8.0.0 (2020-10-11)
 
 ### Bug Fixes
 
@@ -32,9 +32,9 @@
 
 ### BREAKING CHANGES
 
-- minimal supported version of *Angular* is `10.0.0`
+- minimal supported version of *Angular* is `12.0.0`
 - minimal supported version of `@jscrpt/common` is `1.2.0`
-- minimal supported version of `@anglr/common` is `8.0.0`
+- minimal supported version of `@anglr/common` is `8.0.0-beta`
 - removed `jquery` as dependency
 - removed `ReportingExceptionHandlerService`, now should be implemented as custom logger *sink*
 - renamed `ReportingExceptionHandlerOptions` to `AnglrExceptionHandlerOptions`